third_party/WebKit/LayoutTests/external/wpt/selection/removeRange-expected.txt

This is a testharness.js-based test.
FAIL removeRange() with Range 0assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 1assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 2assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 3assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 4assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 5assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 6assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 7assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 14assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 15assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 16assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 17assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 18assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 22assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 23assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 26assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 27assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 28assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 29assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 30assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 31assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 32assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 34assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 35assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 36assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 37assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 38assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
FAIL removeRange() with Range 39assert_throws: Removing a different range should throw function "function () { selection.removeRange(equivalentRange) }" did not throw
PASS removeRange() argument is non-optional Range
Harness: the test ran to completion.
